--- a/Makefile
+++ b/Makefile
@@ -1,8 +1,10 @@
-build:
+build: tsne/bh_sne.pyx \
+	   tsne/bh_sne_src/tsne.cpp \
+	   tsne/bh_sne_src/sptree.cpp \
+	   $(wildcard tsne/bh_sne_src/*.h)
 	python setup.py build_ext --inplace
 
-install:
-	python setup.py build_ext --inplace
+install: build
 	python setup.py install
 
 sdist: