--- a/tsne/bh_sne_src/.clang-format
+++ b/tsne/bh_sne_src/.clang-format
@@ -1,0 +1,3 @@
+BasedOnStyle: Google
+IndentWidth: 4
+AllowShortIfStatementsOnASingleLine: false