--- a/tsne/bh_sne_src/Makefile
+++ b/tsne/bh_sne_src/Makefile
@@ -8,25 +8,19 @@ CFLAGS += -std=c++11 -ffast-math -O3 -flto
 CFLAGS += -ffunction-sections
 LDFLAGS += -Wl,--gc-sections
 
-all: bh_tsne bh_tsne_3d
+all: bh_tsne
 
 bh_tsne: main.o tsne.o sptree.o
-	$(CXX) $(CFLAGS) $(LDFLAGS) $^ -o $@
-
-bh_tsne_3d: tsne_3d.o sptree.o main.o
 	$(CXX) $(CFLAGS) $(LDFLAGS) $^ -o $@
 
 sptree.o: sptree.cpp sptree.h
 	$(CXX) $(CFLAGS) -c $< -o $@
 
-tsne.o: tsne.cpp tsne.h sptree.h vptree.h
+main.o tsne.o: tsne.cpp tsne.h sptree.h vptree.h
 	$(CXX) $(CFLAGS) -c $< -o $@
-
-tsne_3d.o: tsne.cpp tsne.h sptree.h vptree.h
-	$(CXX) $(CFLAGS) -DTSNE3D -c $< -o $@
 
 main.o: main.cpp tsne.h sptree.h vptree.h
 	$(CXX) $(CFLAGS) -c $< -o $@
 
 clean:
-	rm -Rf *.o bh_tsne bh_tsne_3d
+	rm -Rf *.o bh_tsne