NOTE: Absolutely DO NOT merge this before modifying the sake configs to pin the revision being built for PD/CS tarballs for older versions.

A whole bunch of changes:
1.  Code cleanups.
  * Use appropriate C++ data structures, rather than manual memory management.
  * clang-format: no more mixed spaces and tabs.
  * namespaces
  * TSNE shouldn't have been a class - it doesn't have members.
2.  Build improvements.
  * Make a single .so with dynamic dispatch, rather than choosing which dimension mode to use at link time.  Much more reliable and debuggable.
  * Enable LTO
  * Get rid of build dependency on blas/atlas.  It wasn't actually being used, and made it much more complex to package the thing.
   * Break tsne's main, load, or main methods into separate files, which don't need to be built when linking the python extension.
3.  Run-time optimizations.
   * Mostly, reducing the size of the tree nodes by through eliminating redundancy and passing data down from above where appropriate.
   * Flow the templating on dimension further up the stack.
   * Less use needless pointer indirection.
   * Small code restructures to allow better loop vectorization by the compiler.
   * Arithmetic simplifications to eliminate redundant computations.

Testing on cellranger sample 74484, there are some small changes in the output.  I'm not sure if this is because I fixed a bug, introduced a bug, or just a result of numerical instability with some of the changes to arithmetic.

Before:
```
Using random seed: 209652396
Using no_dims = 2, perplexity = 30.000000, and theta = 0.500000
Using max_iter = 1000, stop_lying_iter = 250, mom_switch_iter = 250
Computing input similarities...
Building tree...
 - point 0 of 17030
 - point 10000 of 17030
Input similarities computed in 2.31 seconds (sparsity = 0.007233)!
Learning embedding...
Iteration 50: error is 103.968878 (50 iterations in 3.29 seconds)
Iteration 100: error is 95.993534 (50 iterations in 2.98 seconds)
Iteration 150: error is 84.868840 (50 iterations in 2.63 seconds)
Iteration 200: error is 82.552828 (50 iterations in 2.58 seconds)
Iteration 250: error is 4.310005 (50 iterations in 2.51 seconds)
Iteration 300: error is 3.307660 (50 iterations in 2.29 seconds)
Iteration 350: error is 2.946275 (50 iterations in 2.27 seconds)
Iteration 400: error is 2.718556 (50 iterations in 2.29 seconds)
Iteration 450: error is 2.556061 (50 iterations in 2.30 seconds)
Iteration 500: error is 2.433838 (50 iterations in 2.31 seconds)
Iteration 550: error is 2.337590 (50 iterations in 2.31 seconds)
Iteration 600: error is 2.260225 (50 iterations in 2.32 seconds)
Iteration 650: error is 2.196263 (50 iterations in 2.32 seconds)
Iteration 700: error is 2.143605 (50 iterations in 2.32 seconds)
Iteration 750: error is 2.098874 (50 iterations in 2.32 seconds)
Iteration 800: error is 2.060644 (50 iterations in 2.32 seconds)
Iteration 850: error is 2.027915 (50 iterations in 2.32 seconds)
Iteration 900: error is 1.999645 (50 iterations in 2.32 seconds)
Iteration 950: error is 1.975346 (50 iterations in 2.32 seconds)
Iteration 999: error is 1.955789 (50 iterations in 2.27 seconds)
Fitting performed in 48.59 seconds.
```
After:
```
Using random seed: 209652396
Using D = 10, no_dims = 2, perplexity = 30.000000, and theta = 0.500000
Using max_iter = 1000, stop_lying_iter = 250, mom_switch_iter = 250
Computing input similarities...
Computing approximate perplexity...
Building tree...
 - point 0 of 17030
 - point 10000 of 17030
Input similarities computed in 2.43 seconds (sparsity = 0.007233)!
Learning embedding...
Iteration 50: error is 103.968878 (50 iterations in 2.09 seconds)
Iteration 100: error is 95.993534 (50 iterations in 1.85 seconds)
Iteration 150: error is 84.868779 (50 iterations in 1.61 seconds)
Iteration 200: error is 82.552551 (50 iterations in 1.58 seconds)
Iteration 250: error is 4.310008 (50 iterations in 1.53 seconds)
Iteration 300: error is 3.307299 (50 iterations in 1.39 seconds)
Iteration 350: error is 2.945971 (50 iterations in 1.38 seconds)
Iteration 400: error is 2.715925 (50 iterations in 1.38 seconds)
Iteration 450: error is 2.553057 (50 iterations in 1.39 seconds)
Iteration 500: error is 2.431074 (50 iterations in 1.39 seconds)
Iteration 550: error is 2.335568 (50 iterations in 1.39 seconds)
Iteration 600: error is 2.258516 (50 iterations in 1.39 seconds)
Iteration 650: error is 2.195174 (50 iterations in 1.39 seconds)
Iteration 700: error is 2.142492 (50 iterations in 1.40 seconds)
Iteration 750: error is 2.097975 (50 iterations in 1.41 seconds)
Iteration 800: error is 2.060355 (50 iterations in 1.41 seconds)
Iteration 850: error is 2.027727 (50 iterations in 1.44 seconds)
Iteration 900: error is 1.999039 (50 iterations in 1.42 seconds)
Iteration 950: error is 1.974375 (50 iterations in 1.42 seconds)
Iteration 999: error is 1.953511 (50 iterations in 1.39 seconds)
Fitting performed in 29.65 seconds.
```