@adam-azarchs 2018-06-27
More specifics on the profiling data:
```
55.27% TSNE::_sptree_internal::SPTreeNode<2>::computeNonEdgeForces
16.57% TSNE::run
3.67%  TSNE::_sptree_internal::SPTreeNode<2>::insert
```
The rest is trivial.  Within `computeNonEdgeForces`, the biggest things are:
```asm
       │         if (cum_size == 0 ||
  0.27 │       mov    0x28(%rdi),%eax
 14.48 │       test   %eax,%eax
       │     ↓ je     cc
  1.21 │       cmp    $0x1,%eax
```
and
```asm
       │             double mult = cum_size * sqdist;
  7.65 │       mov    %eax,%eax
```
(which is why the fast-path escape for `cum_size==1`)

The `TSNE::run` samples are actually in `SPTree<2>::computeEdgeForces`, which was inlined, and is as one would expect dominated by the inner loop:
```asm
       │                  const double* data_2 = data + col_P[i] * NDims;
  2.29 │3ab0:   mov    0x148(%rsp),%rsi
  5.82 │        mov    (%r15,%rdi,4),%ebx
  0.65 │        add    %ebx,%ebx
       │                      diffs[d] = data_1[d] - data_2[d];
  0.95 │        movupd (%rax),%xmm1
  2.71 │        movupd (%rsi,%rbx,8),%xmm2
  6.35 │        subpd  %xmm2,%xmm1
       │                      sqdist += diffs[d] * diffs[d];
  0.79 │        movapd %xmm1,%xmm2
  1.00 │        mulpd  %xmm2,%xmm2
  2.94 │        movapd %xmm2,%xmm3
  6.62 │        addsd  %xmm4,%xmm3
  0.51 │        movhlp %xmm2,%xmm2
  1.17 │        addsd  %xmm3,%xmm2
       │                  sqdist = val_P[i] / sqdist;
  2.73 │        movsd  (%r12,%rdi,8),%xmm3
  8.92 │        divsd  %xmm2,%xmm3
       │                      pos_f[d] += sqdist * diffs[d];
  5.33 │        movddu %xmm3,%xmm2
  1.75 │        mulpd  %xmm1,%xmm2
  8.05 │        addpd  %xmm2,%xmm0
 20.55 │        movupd %xmm0,(%rcx)
```
As you can see the for loops don't show up because they've been unrolled and vectorized, which is of course the whole point of templating on `NDims`.