@adam-azarchs 2019-03-29
Closing in favor of #10.  Though this does more optimization than that one, that one made its changes more carefully.  Keeping this branch around however for future reference.